Refactor bulk download class to share common functionality with activity streams handler in the future
