+++ src/main/java/edu/cornell/library/integration/authority/activitystreams/Utils.java
@@ -1,0 +1,337 @@
+package edu.cornell.library.integration.authority.activitystreams;
+
+import java.io.IOException;
+import java.io.InputStream;
+import java.io.PrintWriter;
+import java.io.StringReader;
+import java.io.StringWriter;
+import java.net.URI;
+import java.net.URISyntaxException;
+import java.net.http.HttpClient;
+import java.net.http.HttpClient.Redirect;
+import java.net.http.HttpRequest;
+import java.net.http.HttpResponse;
+import java.net.http.HttpResponse.BodyHandler;
+import java.sql.Connection;
+import java.sql.PreparedStatement;
+import java.sql.SQLException;
+import java.sql.Statement;
+import java.time.LocalDateTime;
+import java.time.format.DateTimeFormatter;
+import java.util.ArrayList;
+import java.util.Arrays;
+import java.util.List;
+import java.util.Optional;
+
+import com.apicatalog.jsonld.JsonLd;
+import com.apicatalog.jsonld.JsonLdError;
+import com.apicatalog.jsonld.JsonLdErrorCode;
+import com.apicatalog.jsonld.document.Document;
+import com.apicatalog.jsonld.document.JsonDocument;
+
+import edu.cornell.library.integration.authority.AuthoritySource;
+import jakarta.json.JsonArray;
+import jakarta.json.JsonObject;
+import jakarta.json.JsonString;
+import jakarta.json.JsonStructure;
+import jakarta.json.JsonValue;
+import jakarta.json.JsonValue.ValueType;
+
+public class Utils {
+	public static final String UPDATE_TABLE = "authorityUpdateActivityStreams";
+	private static final HttpClient HTTP_CLIENT = HttpClient.newBuilder().followRedirects(Redirect.NORMAL).build();
+
+	public static void addBatch(PreparedStatement insertStmt, AuthorityParsedData data, String addedDate) throws SQLException {
+		insertStmt.setString(1, data.id);
+		insertStmt.setString(2, data.lccn);
+		insertStmt.setInt(3, data.vocab.ordinal());
+		insertStmt.setInt(4, data.recordStatus.ordinal());
+		insertStmt.setString(5, data.authorativeLabel);
+		insertStmt.setInt(6, data.headingType.ordinal());
+		insertStmt.setBoolean(7, data.isSubdivision);
+		insertStmt.setBoolean(8, data.undifferentiated);
+		insertStmt.setString(9, data.moddate);
+		insertStmt.setString(10, addedDate);
+		insertStmt.setInt(11, data.numUpdates);
+		insertStmt.setString(12, data.source);
+		insertStmt.addBatch();
+	}
+
+	public static int countNumUpdate(JsonArray graph, JsonObject mainEntry) {
+		JsonValue adminMds = mainEntry.get("madsrdf:adminMetadata");
+		List<String> adminMdIds = Utils.getIdAsList(adminMds);
+		return adminMdIds.size();
+	}
+
+	public static AuthoritySource extractVocab(String id) {
+		if (id.startsWith("gf")) return AuthoritySource.LCGFT;
+		if (id.startsWith("sj")) return AuthoritySource.LCJSH;
+		if (id.startsWith("sh")) return AuthoritySource.LCSH;
+		if (id.startsWith("n"))  return AuthoritySource.NAF;
+		return AuthoritySource.UNK;
+	}
+
+	/*
+	 * For madsrdf:authoritativeLabel
+	 * 1. if it is null, return null (happens for deprecated item)
+	 * 2. if it is a string, return it
+	 * 3. if it is a list, return the first string representation
+	 *    if string representation is not found on the list, return @value from last object representation
+	 * 4. if it is an object (map), return the @value
+	 * If above doesn't resolve to anything, return null
+	 */
+	public static String getAuthorativeLabel(JsonObject mainEntry) {
+		JsonValue authLabel = mainEntry.get("madsrdf:authoritativeLabel");
+		if (authLabel == null)
+			return null;
+		if (authLabel.getValueType() == ValueType.STRING)
+			return Utils.getString(authLabel);
+		String labelForLang = null;
+		if (authLabel.getValueType() == ValueType.ARRAY) {
+			for (JsonValue label : authLabel.asJsonArray()) {
+				if (label.getValueType() == ValueType.STRING)
+					return Utils.getString(label);
+				else if (label.getValueType() == ValueType.OBJECT)
+					labelForLang = label.asJsonObject().getJsonString("@value").getString();
+			}
+		} else if (authLabel.getValueType() == ValueType.OBJECT)
+			return authLabel.asJsonObject().getString("@value");
+		return labelForLang;
+	}
+
+	public static String getDestName(String url) throws URISyntaxException {
+		URI uri = new URI(url);
+		String path = uri.getPath();
+		String basename = path.substring(path.lastIndexOf('/') + 1);
+		if (!basename.endsWith(".gz"))
+			throw new IllegalArgumentException("Input is not gzip! " + url);
+
+		return basename.substring(0, basename.length() - 3);
+	}
+
+	public static JsonObject getFramedDoc(String context, InputStream jsonld) throws InterruptedException, IOException, JsonLdError {
+		Document doc = JsonDocument.of(jsonld);
+		return JsonLd.compact(doc, context).get();
+	}
+
+	public static List<String> getIdAsList(JsonValue arg) {
+		List<String> ids = new ArrayList<>();
+		if (arg.getValueType() == ValueType.ARRAY) {
+			for (JsonValue val : arg.asJsonArray())
+				ids.add(val.asJsonObject().getJsonString("@id").getString());
+		} else if (arg.getValueType() == ValueType.OBJECT)
+			ids.add(arg.asJsonObject().getJsonString("@id").getString());
+		return ids;
+	}
+
+	public static JsonObject getJsonObjectForId(JsonArray graph, String id) {
+		for (JsonValue entry : graph) {
+			JsonObject obj = entry.asJsonObject();
+			String entryId = obj.getString("@id");
+			if (id.equalsIgnoreCase(entryId))
+				return obj;
+		}
+		return JsonValue.EMPTY_JSON_OBJECT;
+	}
+
+	public static JsonObject getLatestRecordInfo(JsonArray graph, JsonObject mainEntry) {
+		JsonValue adminMds = mainEntry.get("madsrdf:adminMetadata");
+		List<String> adminMdIds = Utils.getIdAsList(adminMds);
+		JsonObject recordInfo = JsonValue.EMPTY_JSON_OBJECT;
+		String latest = null;
+		for (String adminMdId : adminMdIds) {
+			JsonObject ri = Utils.getJsonObjectForId(graph, adminMdId);
+			if (latest == null) {
+				recordInfo = ri;
+				latest = recordInfo.getJsonObject("ri:recordChangeDate").getJsonString("@value").getString();
+			} else {
+				String changeDate = ri.getJsonObject("ri:recordChangeDate").getJsonString("@value").getString();
+				if (changeDate.compareToIgnoreCase(latest) > 0) {
+					recordInfo = ri;
+					latest = changeDate;
+				}
+			}
+		}
+		return recordInfo;
+	}
+
+	public static String getStackTraceAsString(Throwable throwable) {
+		StringWriter sw = new StringWriter();
+		PrintWriter pw = new PrintWriter(sw);
+		throwable.printStackTrace(pw);
+		return sw.toString();
+	}
+
+	// Get string value or null if key doesn't exist or not a string type
+	public static String getString(JsonObject obj, String key) {
+		JsonValue value = obj.get(key);
+		if (value == null)
+			return null;
+		if (value.getValueType() == JsonValue.ValueType.STRING)
+			return ((JsonString) value).getString();
+		return null;
+	}
+
+	/*
+	 * Shorthand to get string value from a generic type JsonValue
+	 * If it is an object, return the @value as string
+	 */
+	public static String getString(JsonValue value) {
+		if (value.getValueType() == ValueType.STRING)
+			return ((JsonString) value).getString();
+		else if (value.getValueType() == ValueType.OBJECT)
+			return value.asJsonObject().getJsonString("@value").getString();
+		return null;
+	}
+
+	public static String getToday() {
+		LocalDateTime now = LocalDateTime.now();
+		DateTimeFormatter formatter = DateTimeFormatter.ofPattern("yyyy-MM-dd");
+		return now.format(formatter);
+	}
+
+	public static <T> HttpResponse<T> httpGet(String url, BodyHandler<T> bodyHandler) throws IOException, InterruptedException {
+		HttpRequest request = HttpRequest.newBuilder()
+				.uri(URI.create(url)).build();
+		return HTTP_CLIENT.send(request, bodyHandler);
+	}
+
+	public static boolean parseIsMemberOfMADSCollection(JsonValue isMemberOf, String idUrl) {
+		if (isMemberOf == null)
+			return false;
+		if (isMemberOf.getValueType() == ValueType.STRING)
+			return idUrl.equalsIgnoreCase(Utils.getString(isMemberOf));
+		else if (isMemberOf.getValueType() == ValueType.OBJECT)
+			return idUrl.equalsIgnoreCase(Utils.getString(isMemberOf.asJsonObject(), "@id"));
+		else if (isMemberOf.getValueType() == ValueType.ARRAY) {
+			for (JsonValue memberOf : isMemberOf.asJsonArray()) {
+				if (idUrl.equalsIgnoreCase(memberOf.asJsonObject().getJsonString("@id").getString()))
+					return true;
+			}
+		}
+		return false;
+	}
+
+	public static boolean parseIsSubdivision(JsonValue isMemberOf) {
+		return parseIsMemberOfMADSCollection(isMemberOf, Constants.SUBDIV_URL);
+	}
+
+	public static JsonObject parseJsonLd(InputStream is) throws JsonLdError {
+		Document document = JsonDocument.of(is);
+		Optional<JsonStructure> jsonContentOptional = document.getJsonContent();
+		if (jsonContentOptional.isPresent()) {
+			JsonStructure jsonContent = jsonContentOptional.get();
+			if (jsonContent.getValueType() == ValueType.OBJECT)
+				return jsonContent.asJsonObject();
+		}
+
+		throw new JsonLdError(JsonLdErrorCode.LOADING_DOCUMENT_FAILED, "Failed to parse jsonld");
+	}
+
+	public static AuthorityParsedData parseAuthorityData(JsonObject doc) throws JsonLdError, URISyntaxException {
+		String docUri = "http://id.loc.gov" + doc.getJsonString("@id").getString();
+		JsonArray graph = doc.getJsonArray("@graph");
+		
+		AuthorityParsedData parsed = new AuthorityParsedData();
+		JsonObject mainEntry = Utils.getJsonObjectForId(graph, docUri);
+
+		parsed.id = mainEntry.getJsonString("@id").getString();
+		parsed.lccn = Utils.getString(mainEntry, "identifiers:lccn");
+
+		String path = new URI(parsed.id).getPath();
+		int lastSlashIndex = path.lastIndexOf('/');
+		parsed.vocab = extractVocab(path.substring(lastSlashIndex + 1));
+
+		JsonValue isMemberOf = mainEntry.get("madsrdf:isMemberOfMADSCollection");
+		parsed.undifferentiated = parseUndifferentiated(isMemberOf);
+		parsed.isSubdivision = parseIsSubdivision(isMemberOf);
+
+		parsed.authorativeLabel = getAuthorativeLabel(mainEntry);
+		// If subdivision, we will try to replace -- to >
+		if (parsed.isSubdivision)
+			parsed.authorativeLabel = parsed.authorativeLabel.replace("--", " > ");
+
+		parsed.headingType = parseHeadingType(mainEntry);
+
+		JsonObject riRecord = getLatestRecordInfo(graph, mainEntry);
+		parsed.moddate = riRecord.getJsonObject("ri:recordChangeDate").getString("@value");
+		String recordStatus = riRecord.getString("ri:recordStatus");
+		parsed.recordStatus = MadsRecordStatus.byName(recordStatus);
+
+		parsed.numUpdates = countNumUpdate(graph, mainEntry);
+
+		parsed.source = doc.toString();
+
+		return parsed;
+	}
+
+	public static AuthorityParsedData parseAuthorityData(String jsonld) throws JsonLdError, URISyntaxException {
+		JsonObject doc = parseJsonLd(jsonld);
+		return parseAuthorityData(doc);
+	}
+
+	public static boolean parseUndifferentiated(JsonValue isMemberOf) {
+		return parseIsMemberOfMADSCollection(isMemberOf, Constants.UNDIFF_URL);
+	}
+
+	public static MadsHeadingType parseHeadingType(JsonObject mainEntry) {
+		JsonValue ht = mainEntry.get("@type");
+		if (ht.getValueType() == ValueType.ARRAY) {
+			for (JsonValue type : ht.asJsonArray()) {
+				MadsHeadingType t = MadsHeadingType.byType(Utils.getString(type));
+				if (t != null)
+					return t;
+			}
+		} else if (ht.getValueType() == ValueType.STRING) {
+			return MadsHeadingType.byType(Utils.getString(ht));
+		}
+		return null;
+	}
+
+	public static JsonObject parseJsonLd(String jsonld) throws JsonLdError {
+		StringReader reader = new StringReader(jsonld);
+		Document document = JsonDocument.of(reader);
+		Optional<JsonStructure> jsonContentOptional = document.getJsonContent();
+		if (jsonContentOptional.isPresent()) {
+			JsonStructure jsonContent = jsonContentOptional.get();
+			if (jsonContent.getValueType() == ValueType.OBJECT)
+				return jsonContent.asJsonObject();
+		}
+
+		throw new JsonLdError(JsonLdErrorCode.LOADING_DOCUMENT_FAILED, "Failed to parse jsonld " + jsonld);
+	}
+
+	public static PreparedStatement replaceStmt(Connection authorityDB) throws SQLException {
+		return authorityDB.prepareStatement(
+				"""
+REPLACE INTO %s (id,lccn,vocabulary,recordStatus,heading,headingType,isSubdivision,undifferentiated,moddate,addedDate,numUpdates,source)
+	VALUES (?,?,?,?,?,?,?,?,?,?,?,?)""".formatted(UPDATE_TABLE));
+	}
+
+	public static void setUpDatabase(Connection authority) throws SQLException {
+		List<String> sqls = Arrays.asList(
+				"""
+CREATE TABLE IF NOT EXISTS `%s` (
+	`id` VARCHAR(255) NOT NULL,
+	`lccn` VARCHAR(50) DEFAULT NULL,
+	`vocabulary` TINYINT(4) DEFAULT NULL,
+	`recordStatus` TINYINT(1) DEFAULT NULL,
+	`heading` TEXT DEFAULT NULL,
+	`headingType` TINYINT(4) DEFAULT NULL,
+	`isSubdivision` TINYINT(1) DEFAULT NULL,
+	`undifferentiated` TINYINT(1) DEFAULT NULL,
+	`moddate` VARCHAR(30) DEFAULT NULL,
+	`addedDate` VARCHAR(30) DEFAULT NULL,
+	`numUpdates` INT DEFAULT 0,
+	`source` JSON NOT NULL,
+	UNIQUE KEY `unique_id` (`id`, `numUpdates`, `moddate`),
+	KEY `idx_id` (`id`),
+	KEY `idx_added_date` (`addedDate`),
+	KEY `idx_moddate` (`moddate`)
+) ENGINE=MyISAM DEFAULT CHARSET=utf8""".formatted(UPDATE_TABLE));
+		try ( Statement stmt = authority.createStatement() ) {
+			for (String sql : sqls)
+				stmt.execute(sql);
+		}
+	}
+}
