+++ src/test/java/edu/cornell/library/integration/authority/activitystreams/BulkDownloadHandlerTest.java
@@ -35,7 +35,7 @@ public class BulkDownloadHandlerTest extends DbBaseTest {
 		String base = Path.of("src", "test", "resources", "edu", "cornell", "library", "integration", "authority").toString();
 		if (useTestContainers != null) {
 			try (Connection authority = config.getDatabaseConnection("Authority")) {
-				Helper.setUpDatabase(authority);
+				Utils.setUpDatabase(authority);
 			}
 		} else if (useSqlite != null) {
 			String extraSqlite = Path.of(base, "authority_extra_sqlite_create.sql").toString();
@@ -70,15 +70,16 @@ public class BulkDownloadHandlerTest extends DbBaseTest {
 		Path tempFile = Files.createTempFile("myPrefix", ".tmp");
 		try (OutputStream out = Files.newOutputStream(tempFile);
 			 Connection authority = config.getDatabaseConnection("Authority");
-			 PreparedStatement stmt = authority.prepareStatement("SELECT * FROM %s WHERE id = ? AND moddate = ?".formatted(Helper.UPDATE_TABLE))) {
+			 PreparedStatement stmt = authority.prepareStatement("SELECT * FROM %s WHERE id = ? AND numUpdates = ? AND moddate = ?".formatted(Utils.UPDATE_TABLE))) {
 			out.write(builder.toString().getBytes(StandardCharsets.UTF_8));
 
-			String today = Helper.getToday();
+			String today = Utils.getToday();
 			BulkDownloadHandler bdh = new BulkDownloadHandler();
 			bdh.processData(today, tempFile, authority, 2);
 
 			stmt.setString(1, "http://id.loc.gov/authorities/names/n00000001");
-			stmt.setString(2, "2025-08-05T02:34:09");
+			stmt.setInt(2, 2);
+			stmt.setString(3, "2025-08-05T02:34:09");
 			ResultSet rs = stmt.executeQuery();
 			assertEquals(true, rs.next());
 			assertEquals("n 00000001", rs.getString("lccn"));
@@ -93,76 +94,5 @@ public class BulkDownloadHandlerTest extends DbBaseTest {
 		} finally {
 			Files.delete(tempFile);
 		}
-	}
-
-	@Test
-	public void parseBulkEntryTest() throws IOException, JsonLdError, URISyntaxException {
-		BulkDownloadHandler lbd = new BulkDownloadHandler();
-		String content = Files.readString(resources.get("names_madsrdf_1.json"));
-		AuthorityParsedData data = lbd.parseBulkEntry(content);
-		assertEquals("http://id.loc.gov/authorities/names/n00000001", data.id);
-		assertEquals("n 00000001", data.lccn);
-		assertEquals(AuthoritySource.valueOf("NAF"), data.vocab);
-		assertEquals(MadsRecordStatus.valueOf("REVISED"), data.recordStatus);
-		assertEquals("McQuerry, Maureen, 1955-", data.authorativeLabel);
-		assertEquals(MadsHeadingType.valueOf("PERSONAL_NAME"), data.headingType);
-		assertEquals(false, data.isSubdivision);
-		assertEquals(false, data.undifferentiated);
-		assertEquals("2025-08-05T02:34:09", data.moddate);
-
-		content = Files.readString(resources.get("names_madsrdf_2.json"));
-		data = lbd.parseBulkEntry(content);
-		assertEquals("Belʹdiĭ, A. I︠A︡.", data.authorativeLabel);
-
-		content = Files.readString(resources.get("subjects_madsrdf_1.json"));
-		data = lbd.parseBulkEntry(content);
-		assertEquals(AuthoritySource.valueOf("LCSH"), data.vocab);
-		assertEquals("ActionScript (Computer program language)", data.authorativeLabel);
-		assertEquals(MadsHeadingType.valueOf("TOPIC"), data.headingType);
-
-		content = Files.readString(resources.get("names_madsrdf_subdivision_1.json"));
-		data = lbd.parseBulkEntry(content);
-		assertEquals(null, data.lccn);
-		assertEquals("Germany > Marienstatt", data.authorativeLabel);
-		assertEquals(MadsHeadingType.valueOf("HIERARCHICAL_GEOGRAPHIC"), data.headingType);
-
-		content = Files.readString(resources.get("subjects_madsrdf_subdivision_1.json"));
-		data = lbd.parseBulkEntry(content);
-		assertEquals("sh 00008433", data.lccn);
-		assertEquals("Reinstatement", data.authorativeLabel);
-		assertEquals(MadsHeadingType.valueOf("TOPIC"), data.headingType);
-		assertEquals(true, data.isSubdivision);
-
-		content = Files.readString(resources.get("subjects_madsrdf_hai_van_pass.json"));
-		data = lbd.parseBulkEntry(content);
-		assertEquals("Hải Vân Pass (Vietnam)", data.authorativeLabel);
-		assertEquals(MadsHeadingType.valueOf("GEOGRAPHIC"), data.headingType);
-		assertEquals(false, data.isSubdivision);
-
-		content = Files.readString(resources.get("names_undifferentiated_1.json"));
-		data = lbd.parseBulkEntry(content);
-		assertEquals("Mason, Jack", data.authorativeLabel);
-		assertEquals(true, data.undifferentiated);
-
-		content = Files.readString(resources.get("names_deprecated_1.json"));
-		data = lbd.parseBulkEntry(content);
-		assertEquals(null, data.lccn);
-		assertEquals(MadsRecordStatus.valueOf("DEPRECATED"), data.recordStatus);
-		assertEquals(null, data.authorativeLabel);
-
-		content = Files.readString(resources.get("names_name_title_1.json"));
-		data = lbd.parseBulkEntry(content);
-		assertEquals(MadsHeadingType.valueOf("NAME_TITLE"), data.headingType);
-
-		// Following two have -- in authoratativeLabel but is not subdivision
-		content = Files.readString(resources.get("subjects_name_title_1.json"));
-		data = lbd.parseBulkEntry(content);
-		assertEquals(MadsHeadingType.valueOf("NAME_TITLE"), data.headingType);
-		assertEquals("Texas. Declaration of Independence--Signers", data.authorativeLabel);
-
-		content = Files.readString(resources.get("subjects_complex_subject_1.json"));
-		data = lbd.parseBulkEntry(content);
-		assertEquals(MadsHeadingType.valueOf("COMPLEX_SUBJECT"), data.headingType);
-		assertEquals("Space vehicles--Doppler tracking", data.authorativeLabel);
 	}
 }
