+++ src/test/java/edu/cornell/library/integration/authority/activitystreams/UtilsTest.java
@@ -1,0 +1,106 @@
+package edu.cornell.library.integration.authority.activitystreams;
+
+import static org.junit.jupiter.api.Assertions.assertEquals;
+
+import java.io.IOException;
+import java.net.URISyntaxException;
+import java.nio.file.Files;
+import java.nio.file.Path;
+import java.util.Arrays;
+import java.util.LinkedHashMap;
+import java.util.Map;
+
+import org.junit.BeforeClass;
+import org.junit.Test;
+
+import com.apicatalog.jsonld.JsonLdError;
+
+import edu.cornell.library.integration.authority.AuthoritySource;
+
+public class UtilsTest {
+	static Map<String, Path> resources = new LinkedHashMap<>();
+
+	@BeforeClass
+	public static void setup() {
+		String base = Path.of("src", "test", "resources", "edu", "cornell", "library", "integration", "authority").toString();
+		for (String file : Arrays.asList("names_madsrdf_1.json", "names_madsrdf_2.json",
+				"subjects_madsrdf_1.json", "names_madsrdf_subdivision_1.json",
+				"subjects_madsrdf_subdivision_1.json", "subjects_madsrdf_hai_van_pass.json",
+				"names_undifferentiated_1.json", "names_deprecated_1.json",
+				"names_name_title_1.json", "subjects_name_title_1.json",
+				"subjects_complex_subject_1.json")) {
+			Path path = Path.of(base, file);
+			resources.put(file, path);
+		}
+	}
+
+	@Test
+	public void parseBulkEntryTest() throws IOException, JsonLdError, URISyntaxException {
+		String content = Files.readString(resources.get("names_madsrdf_1.json"));
+		AuthorityParsedData data = Utils.parseAuthorityData(content);
+		assertEquals("http://id.loc.gov/authorities/names/n00000001", data.id);
+		assertEquals("n 00000001", data.lccn);
+		assertEquals(AuthoritySource.valueOf("NAF"), data.vocab);
+		assertEquals(MadsRecordStatus.valueOf("REVISED"), data.recordStatus);
+		assertEquals("McQuerry, Maureen, 1955-", data.authorativeLabel);
+		assertEquals(MadsHeadingType.valueOf("PERSONAL_NAME"), data.headingType);
+		assertEquals(false, data.isSubdivision);
+		assertEquals(false, data.undifferentiated);
+		assertEquals("2025-08-05T02:34:09", data.moddate);
+
+		content = Files.readString(resources.get("names_madsrdf_2.json"));
+		data = Utils.parseAuthorityData(content);
+		assertEquals("Belʹdiĭ, A. I︠A︡.", data.authorativeLabel);
+
+		content = Files.readString(resources.get("subjects_madsrdf_1.json"));
+		data = Utils.parseAuthorityData(content);
+		assertEquals(AuthoritySource.valueOf("LCSH"), data.vocab);
+		assertEquals("ActionScript (Computer program language)", data.authorativeLabel);
+		assertEquals(MadsHeadingType.valueOf("TOPIC"), data.headingType);
+
+		content = Files.readString(resources.get("names_madsrdf_subdivision_1.json"));
+		data = Utils.parseAuthorityData(content);
+		assertEquals(null, data.lccn);
+		assertEquals("Germany > Marienstatt", data.authorativeLabel);
+		assertEquals(MadsHeadingType.valueOf("HIERARCHICAL_GEOGRAPHIC"), data.headingType);
+
+		content = Files.readString(resources.get("subjects_madsrdf_subdivision_1.json"));
+		data = Utils.parseAuthorityData(content);
+		assertEquals("sh 00008433", data.lccn);
+		assertEquals("Reinstatement", data.authorativeLabel);
+		assertEquals(MadsHeadingType.valueOf("TOPIC"), data.headingType);
+		assertEquals(true, data.isSubdivision);
+
+		content = Files.readString(resources.get("subjects_madsrdf_hai_van_pass.json"));
+		data = Utils.parseAuthorityData(content);
+		assertEquals("Hải Vân Pass (Vietnam)", data.authorativeLabel);
+		assertEquals(MadsHeadingType.valueOf("GEOGRAPHIC"), data.headingType);
+		assertEquals(false, data.isSubdivision);
+
+		content = Files.readString(resources.get("names_undifferentiated_1.json"));
+		data = Utils.parseAuthorityData(content);
+		assertEquals("Mason, Jack", data.authorativeLabel);
+		assertEquals(true, data.undifferentiated);
+
+		content = Files.readString(resources.get("names_deprecated_1.json"));
+		data = Utils.parseAuthorityData(content);
+		assertEquals(null, data.lccn);
+		assertEquals(MadsRecordStatus.valueOf("DEPRECATED"), data.recordStatus);
+		assertEquals(null, data.authorativeLabel);
+
+		content = Files.readString(resources.get("names_name_title_1.json"));
+		data = Utils.parseAuthorityData(content);
+		assertEquals(MadsHeadingType.valueOf("NAME_TITLE"), data.headingType);
+
+		// Following two have -- in authoratativeLabel but is not subdivision
+		content = Files.readString(resources.get("subjects_name_title_1.json"));
+		data = Utils.parseAuthorityData(content);
+		assertEquals(MadsHeadingType.valueOf("NAME_TITLE"), data.headingType);
+		assertEquals("Texas. Declaration of Independence--Signers", data.authorativeLabel);
+
+		content = Files.readString(resources.get("subjects_complex_subject_1.json"));
+		data = Utils.parseAuthorityData(content);
+		assertEquals(MadsHeadingType.valueOf("COMPLEX_SUBJECT"), data.headingType);
+		assertEquals("Space vehicles--Doppler tracking", data.authorativeLabel);
+	}
+}
