+++ src/test/resources/edu/cornell/library/integration/authority/authority_extra_sqlite_create.sql
@@ -1,1 +1,1 @@
-CREATE TABLE IF NOT EXISTS `authorityUpdateActivityStreams` (`id` text DEFAULT NULL, `lccn` text DEFAULT NULL, `vocabulary` int DEFAULT NULL, `recordStatus` int DEFAULT NULL, `heading` text DEFAULT NULL, `headingType` int DEFAULT NULL, `isSubdivision` int DEFAULT NULL, `undifferentiated` int DEFAULT NULL, `moddate` text DEFAULT NULL, `addedDate` text DEFAULT NULL, `numUpdates` int DEFAULT 0, source text NOT NULL, PRIMARY KEY (`id`, `numUpdates`));
+CREATE TABLE IF NOT EXISTS `authorityUpdateActivityStreams` (`id` text DEFAULT NULL, `lccn` text DEFAULT NULL, `vocabulary` int DEFAULT NULL, `recordStatus` int DEFAULT NULL, `heading` text DEFAULT NULL, `headingType` int DEFAULT NULL, `isSubdivision` int DEFAULT NULL, `undifferentiated` int DEFAULT NULL, `moddate` text DEFAULT NULL, `addedDate` text DEFAULT NULL, `numUpdates` int DEFAULT 0, source text NOT NULL, PRIMARY KEY (`id`, `numUpdates`, `moddate`));
